+++ modules/mrpt_maps/tests/COctoMap_unittest.cpp
@@ -43,12 +43,14 @@ TEST(COctoMapTests, updateVoxels)
   mrpt::math::TPoint3D pt;
 
   pt = mrpt::math::TPoint3D(1, 1, 1);
-  is_mapped = map.getPointOccupancy(pt.x, pt.y, pt.z, occup);
+  is_mapped = map.getPointOccupancy(
+      static_cast<float>(pt.x), static_cast<float>(pt.y), static_cast<float>(pt.z), occup);
   EXPECT_GT(occup, 0.5);
   EXPECT_TRUE(is_mapped);
 
   pt = mrpt::math::TPoint3D(-1, -1, 1);
-  is_mapped = map.getPointOccupancy(pt.x, pt.y, pt.z, occup);
+  is_mapped = map.getPointOccupancy(
+      static_cast<float>(pt.x), static_cast<float>(pt.y), static_cast<float>(pt.z), occup);
   EXPECT_LT(occup, 0.5);
   EXPECT_TRUE(is_mapped);
 }
